requirements.txt
Flask
google-generativeai
Flask-Cors
openai